pyproject.toml
[tool.poetry]
name = "trainstation-workload-simulation"
version = "0.1.0"
description = ""
authors = ["kaponda mulambia <44966170+kmulambia@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pika = "^1.3.2"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
